requirements.txt
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
numpy==1.23.1
pandas==1.4.3
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1